requirements.txt
numpy==2.3.5
opencv-python==4.12.0.88
torch==2.5.1
torchvision==0.20.1
PyQt5==5.15.11
SQLAlchemy==2.0.43
pillow==10.4.0
python-dotenv==1.0.1
PyMySQL==1.1.1
